move LanguageService back to lib.rs
